Move osprepare from ciao
Fixes: #110
