+++ osprepare/distro.go
