+++ osprepare/distro_test.go
