+++ osprepare/doc.go
