+++ osprepare/os_release.go
