+++ osprepare/os_release_test.go
